working on security dev

+++ public/assignment/server/services/user.service.server.js
@@ -1,18 +1,66 @@
 "use strict";
 
+var passport = require('passport');
+var LocalStrategy = require('passport-local').Strategy;
+//var mongoose      = require("mongoose");
+
 module.exports = function (app, formModel, userModel) {
-    app.post("/api/assignment/user", addUser);
+    app.post("/api/assignment/user", admin, addUser);
+    app.post("/api/assignment/register", register);
+    app.post("/api/assignment/login", passport.authenticate('local'), login);
     app.get("/api/assignment/user", getUser);
+    app.post("/api/assignment/logout", logout);
+    app.get("/api/assignment/loggedin", loggedin);
     app.get("/api/assignment/user/:id", profile);
     app.put("/api/assignment/user/:id", updateUser);
-    app.delete("/api/assignment/user/:id", deleteUser);
+    app.delete("/api/assignment/user/:id", admin, deleteUser);
+
+    passport.use(new LocalStrategy(localStrategy));
+    passport.serializeUser(serializeUser);
+    passport.deserializeUser(deserializeUser);
+
+    function localStrategy(username, password, done) {
+        userModel
+            .findUserByCredentials({username: username, password: password})
+            .then(
+                function (user) {
+                    if (!user) {
+                        return done(null, false);
+                    }
+                    return done(null, user);
+                },
+                function (err) {
+                    if (err) {
+                        return done(err);
+                    }
+                }
+            );
+    }
+
+    function serializeUser(user, done) {
+        done(null, user);
+    }
+
+    function deserializeUser(user, done) {
+        userModel
+            .findUserById(user._id)
+            .then(
+                function (user) {
+                    done(null, user);
+                },
+                function (err) {
+                    done(err, null);
+                }
+            );
+    }
+
+    function loggedin(req, res) {
+        res.send(req.isAuthenticated() ? req.user : '0');
+    }
 
     function getUser(req, res) {
         var username = req.query.username;
-        var password = req.query.password;
-        if (username != null && password != null) {
-            login(req, res);
-        } else if (username != null) {
+        if (username != null) {
             findUserByUsername(req, res);
         } else {
             all(req, res);
@@ -20,15 +68,19 @@ module.exports = function (app, formModel, userModel) {
     }
 
     function all(req, res) {
-        userModel.findAllUsers()
-            .then(
-                function (doc) {
-                    res.json(doc);
-                },
-                function (err) {
-                    res.status(400).send(err);
-                }
-            );
+        if (isAdmin(req.user)) {
+            userModel.findAllUsers()
+                .then(
+                    function (doc) {
+                        res.json(doc);
+                    },
+                    function (err) {
+                        res.status(400).send(err);
+                    }
+                );
+        } else {
+            res.status(403);
+        }
     }
 
     function profile(req, res) {
@@ -58,19 +110,51 @@ module.exports = function (app, formModel, userModel) {
             );
     }
 
+    function register(req, res) {
+        var newUser = req.body;
+        newUser.roles = ['student'];
+
+        userModel
+            .findUserByUsername(newUser.username)
+            .then(
+                function (user) {
+                    if (user) {
+                        res.json(null);
+                    } else {
+                        return userModel.createUser(newUser);
+                    }
+                },
+                function (err) {
+                    res.status(400).send(err);
+                }
+            )
+            .then(
+                function (user) {
+                    if (user) {
+                        req.login(user, function (err) {
+                            if (err) {
+                                res.status(400).send(err);
+                            } else {
+                                res.json(user);
+                            }
+                        });
+                    }
+                },
+                function (err) {
+                    res.status(400).send(err);
+                }
+            );
+    }
+
     function addUser(req, res) {
         var user = req.body;
 
         userModel.createUser(user)
             // handle model promise
             .then(
-                // login user if promise resolved
                 function (doc) {
-                    //req.session.currentUser = doc;
-                    //console.log(doc);
                     res.json(doc);
                 },
-                // send error if promise rejected
                 function (err) {
                     res.status(400).send(err);
                 }
@@ -78,20 +162,13 @@ module.exports = function (app, formModel, userModel) {
     }
 
     function login(req, res) {
-        var credentials = {};
-        credentials.username = req.query.username;
-        credentials.password = req.query.password;
-        userModel.findUserByCredentials(credentials)
-            .then(
-                function (doc) {
-                    //req.session.currentUser = doc;
-                    res.json(doc);
-                },
-                // send error if promise rejected
-                function (err) {
-                    res.status(400).send(err);
-                }
-            )
+        var user = req.user;
+        res.json(user);
+    }
+
+    function logout(req, res) {
+        req.logout();
+        res.send(200);
     }
 
     function updateUser(req, res) {
@@ -109,15 +186,31 @@ module.exports = function (app, formModel, userModel) {
     }
 
     function deleteUser(req, res) {
-        var id = req.params.id;
-        userModel.deleteUser(id)
-            .then(
-                function (doc) {
-                    res.json(doc);
-                },
-                function (err) {
-                    res.status(400).send(err);
-                }
-            )
+        if (isAdmin(req.user)) {
+            var id = req.params.id;
+            userModel.deleteUser(id)
+                .then(
+                    function (doc) {
+                        res.json(doc);
+                    },
+                    function (err) {
+                        res.status(400).send(err);
+                    }
+                )
+        } else {
+            res.status(403);
+        }
+    }
+
+    function isAdmin(user) {
+        return user.roles.indexOf("admin") > 0;
+    }
+
+    function authorized(req, res, next) {
+        if (!req.isAuthenticated()) {
+            res.send(401);
+        } else {
+            next();
+        }
     }
 };
